add json server + update page List + rename palyPage
commit
